requirements.txt
fastapi
uvicorn
openai>=1.30
python-dotenv
requests
